feat: Internal States with Button feat: Remount button and event handler feat: FavIcons and manifest reafactor: MobX Componets in own folder reafactor: Content component outside the App refactor: CSS Styles
