+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "react-cra-no-eject-mobx-with-decorators-and-hooks",
   "description": "Create React App with MobX, decorators, hooks, setState() and without eject! Impossible? Yeah yeah yeah :)",
-  "version": "0.3.9",
+  "version": "0.5.11",
   "private": true,
   "author": "Anton Karpenko <i@karpolan.com> (https://karpolan.com)",
   "keywords": [
@@ -25,7 +25,7 @@
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
     "babel-plugin-import": "^1.13.0",
-    "mobx": "^5.15.4",
+    "mobx": "^5.15.5",
     "mobx-react": "^6.2.5",
     "react": "^16.13.1",
     "react-app-rewire-mobx": "^1.0.9",
