+++ public/manifest.json
@@ -1,25 +1,46 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "name": "Create React App with MobX Demo.",
+  "short_name": "React App with MobX",
   "icons": [
     {
-      "src": "favicon.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/x-icon"
+      "src": "/favicon/android-icon-36x36.png",
+      "sizes": "36x36",
+      "type": "image/png",
+      "density": "0.75"
     },
     {
-      "src": "logo192.png",
+      "src": "/favicon/android-icon-48x48.png",
+      "sizes": "48x48",
       "type": "image/png",
-      "sizes": "192x192"
+      "density": "1.0"
     },
     {
-      "src": "logo512.png",
+      "src": "/favicon/android-icon-72x72.png",
+      "sizes": "72x72",
       "type": "image/png",
-      "sizes": "512x512"
+      "density": "1.5"
+    },
+    {
+      "src": "/favicon/android-icon-96x96.png",
+      "sizes": "96x96",
+      "type": "image/png",
+      "density": "2.0"
+    },
+    {
+      "src": "/favicon/android-icon-144x144.png",
+      "sizes": "144x144",
+      "type": "image/png",
+      "density": "3.0"
+    },
+    {
+      "src": "/favicon/android-icon-192x192.png",
+      "sizes": "192x192",
+      "type": "image/png",
+      "density": "4.0"
     }
   ],
   "start_url": ".",
   "display": "standalone",
-  "theme_color": "#000000",
+  "theme_color": "#ffffff",
   "background_color": "#ffffff"
 }
