+++ src/App.css
@@ -67,8 +67,21 @@ main {
   border-width: 1px; */
 }
 
-.component div {
+.component .container {
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+  justify-content: space-evenly;
+}
+
+.component .item {
+  flex: 1;
   padding: 0 8px;
+}
+
+.component .item button {
+  display: block;
+  margin: auto;
 }
 
 .component textarea {
