+++ src/App.js
@@ -1,33 +1,23 @@
-import React from 'react';
+import React, { useState, useCallback } from 'react';
 import { Store } from './store';
-import {
-  Header,
-  Footer,
-  ClassComponentWithDecorator,
-  ClassComponentWithObserver,
-  FunctionComponentWithObserver,
-  FunctionComponentUseObserver,
-  UsingObserverComponent,
-  UsingObserverComponentAsRenderProps,
-} from './components';
+import { Header, Content, Footer } from './components';
 import './App.css';
 
 /**
  * App with MobX store(s) injected via React Context API
  */
 function App() {
+  const [reloadKey, setReloadKey] = useState(0);
+
+  const handleReloadClick = useCallback(() => {
+    setReloadKey(reloadKey + 1);
+  }, [reloadKey]);
+
   return (
     <Store>
       <div className="app">
-        <Header />
-        <main>
-          <ClassComponentWithDecorator />
-          <ClassComponentWithObserver />
-          <FunctionComponentUseObserver />
-          <FunctionComponentWithObserver />
-          <UsingObserverComponent />
-          <UsingObserverComponentAsRenderProps />
-        </main>
+        <Header onRemountClick={handleReloadClick} />
+        <Content key={reloadKey} /* Changing of the key property unmount and mount again the React component */ />
         <Footer />
       </div>
     </Store>
