+++ src/components/MobXObserverComponents/index.js
@@ -1,0 +1,15 @@
+import ClassComponentWithDecorator from './ClassComponentWithDecorator';
+import ClassComponentWithObserver from './ClassComponentWithObserver';
+import FunctionComponentWithObserver from './FunctionComponentWithObserver';
+import FunctionComponentUseObserver from './FunctionComponentUseObserver';
+import UsingObserverComponent from './UsingObserverComponent';
+import UsingObserverComponentAsRenderProps from './UsingObserverComponentAsRenderProps';
+
+export {
+  ClassComponentWithDecorator,
+  ClassComponentWithObserver,
+  FunctionComponentWithObserver,
+  FunctionComponentUseObserver,
+  UsingObserverComponent,
+  UsingObserverComponentAsRenderProps,
+};
