+++ src/components/index.js
@@ -1,19 +1,5 @@
 import Header from './Header';
+import Content from './Content';
 import Footer from './Footer';
-import ClassComponentWithDecorator from './ClassComponentWithDecorator';
-import ClassComponentWithObserver from './ClassComponentWithObserver';
-import FunctionComponentWithObserver from './FunctionComponentWithObserver';
-import FunctionComponentUseObserver from './FunctionComponentUseObserver';
-import UsingObserverComponent from './UsingObserverComponent';
-import UsingObserverComponentAsRenderProps from './UsingObserverComponentAsRenderProps';
 
-export {
-  Header,
-  Footer,
-  ClassComponentWithDecorator,
-  ClassComponentWithObserver,
-  FunctionComponentWithObserver,
-  FunctionComponentUseObserver,
-  UsingObserverComponent,
-  UsingObserverComponentAsRenderProps,
-};
+export { Header, Content, Footer };
